+++ src/pages/list/List.jsx
@@ -1,4 +1,5 @@
 import DonationPage from "./donation/components/DonationPage";
+import ListChart from "./charts/components/ListChart";
 import { createGlobalStyle } from "styled-components";
 import colors from "../../utils/colors";
 window.localStorage.setItem("credit", 30000);
@@ -18,6 +19,7 @@ function List() {
     <>
       <GlobalStyle />
       <DonationPage />
+      <ListChart />
     </>
   );
 }
